refactor: move some files, and ensure correct height
