+++ packages/dm-core/src/index.tsx
@@ -28,3 +28,4 @@ export * from './utils/objectUtilities'
 export * from './utils/truncatePathString'
 export * from './utils/uuid'
 export * from './utils/variables'
+export * from './utils/stringUtilities'
