+++ packages/dm-core/src/utils/stringUtilities.ts
@@ -14,3 +14,15 @@ export const splitString = (
   if (arr.slice(maxsplit).length == 0) return arr
   return [...arr.slice(0, maxsplit), arr.slice(maxsplit).join(separator)]
 }
+
+export function formatBytes(bytes: any, decimals = 2) {
+  if (bytes === 0) return '0 Bytes'
+
+  const k = 1024
+  const dm = decimals < 0 ? 0 : decimals
+  const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB']
+
+  const i = Math.floor(Math.log(bytes) / Math.log(k))
+
+  return parseFloat((bytes / Math.pow(k, i)).toFixed(dm)) + ' ' + sizes[i]
+}
